Renamed PicklistMenu to PicklistSize (#1wut88u)
